requirements.txt
Django==2.1
DjangoRestFramework==3.9.4
django-filter==2.1.0
djangorestframework-oauth==1.1.0
django-oauth-toolkit==1.2.0
pillow==6.0.0
mysql-connector==2.2.9
mysqlclient==1.4.2